// src/pages/games/d2/data/types.ts
export enum Slot {
  Weapon = "Weapon",
  Armor = "Armor",
  Helm = "Helm",
  Shield = "Shield",
}

export enum ItemType {
  Weapon = "Weapon",
  MeleeWeapon = "Melee Weapon",
  Axe = "Axe",
  Mace = "Mace",
  Sword = "Sword",
  Stave = "Stave",
  Club = "Club",
  Polearm = "Polearm",
  Scepter = "Scepter",
  Hammer = "Hammer",
  MissileWeapon = "Missile Weapon",
  Crossbow = "Crossbow",
  Bows = "Bow",
  Javelins = "Javelin",
  Daggers = "Dagger",
  Spear = "Spear",
  Throwing = "Throwing",
  Wand = "Wand",
  AmazonBow = "Amazon Bow",
  AmazonSpear = "AMAZON_Spear",
  SorceressOrb = "Sorceress Orb",
  Claw = "Claw",
  Armor = "Armor",
  BodyArmor = "Body Armor",
  Shield = "Shield",
  Helm = "Helm",
  PaladinShield = "Paladin Shield",
  NecromancerShrunkenHead = "Necromancer Shrunken Head",
  BarbarianHelm = "Barbarian Helm",
  DruidPelt = "Druid Pelt",
}